+++ bindings/gms2/datafiles/shaders/fbo-fp.glsl
@@ -1,0 +1,6 @@
+uniform sampler2D fboColor;
+
+void main(void)
+{
+    gl_FragColor = texture2D(fboColor, gl_TexCoord[0].st);
+}