+++ CHANGELOG.md
@@ -7,10 +7,12 @@ v4.0.0 (by Gecko) - ??/??/????
 - UTF-8 support for file names
 - `Update` is deprecated, use `EngineUpdate` instead
 - `TrisRendered` was removed
+- New function `ViewerRenderObject`
 - `ViewerGetPickedObjectsList` now requires user-provied PickList
 - New functions `PickListCreate`, `PickListClear`, `PickListGetCount`, `PickListGetHit`
 - New constants for `ViewerSetAntiAliasing`: `aa6x = 6`, `aa8x = 7`, `aa16x = 8`, `csa8x = 9`, `csa8xHQ = 10`, `csa16x = 11`, `csa16xHQ = 12`
 - `CameraZoomAll` now requires a Viewer as a second parameter
+- New function `MaterialLibraryGetTextureByName`
 - `SpriteNoZWrite` was removed, use `MaterialSetDepthWrite` instead
 - `MaterialCullFrontFaces` was removed
 - `MaterialSetZWrite` is deprecated, use `MaterialSetDepthWrite` instead
@@ -23,7 +25,8 @@ v4.0.0 (by Gecko) - ??/??/????
 - New function `BaseMeshBuildSilhouetteConnectivityData`
 - New function `ActorSetReference`
 - New functions `ObjectNotifyChange`, `ObjectStructureChanged`, `ObjectClearStructureChanged`
-- New functions `ObjectListCreate`, `ObjectListAdd`, `ObjectListGetCount`.
+- New functions `ObjectListCreate`, `ObjectListAdd`, `ObjectListGetCount`
+- FBO system was redesigned. `FBOSetViewer`, `FBORenderObject`, `FBORenderObjectEx`, `FBOSetOverrideMaterial`, `FBOSetColorTextureFormat` were removed. New functions `FBOSetActive`, `FBOSetAspect`, `FBOSetPickableTarget`, `FBOSetSize`, `FBOSetRootObject`, `FBOSetBackgroundColor`, `FBOSetEnabledRenderBuffers`, `FBOSetSceneScaleFactor`, `FBOSetTargetVisibility`, `FBOSetMaterialLibrary`, `FBOSetColorTextureName`, `FBOSetDepthTextureName`, `FBOSetClearOptions`, `FBOSetStencilPrecision`.
 
 v3.9.2 (by Gecko) - 09/04/2022
 ------------------------------